Create Project3
Login capability for employees.
Add new user.
Save users to file.
Save password as hash
Display the inventory of goods.
Issuing an invoice to the customer.
Save each invoice as a csv file
The goods that are issued in the invoice must be subtracted from the inventory.
Save the history of invoices in the file.
Save inventory of goods in the file.
The ability to use different currencies such as euros and dollars and convert them to each other.
Ability to update exchange rates manually.
Update exchange rate automatically API.